refactor: split identifier case plan state helpers
